requirements.txt
APScheduler==3.9.1
beautifulsoup4==4.11.1
Flask==2.0.2
Pillow==9.1.0
pt_core_news_sm==3.2.0
requests==2.26.0
ru_core_news_sm==3.2.0
spacy==3.2.2
SQLAlchemy==1.4.31
